Fix narrowing error
Without this the following error could appear:
```
[ 59%] Building CXX object CMakeFiles/OpenNFS.dir/src/Camera/HermiteCamera.cpp.o
/home/wouter/Personal/OpenNFS/src/Camera/HermiteCamera.cpp:13:26: error: non-constant-expression cannot be narrowed from type 'double' to 'float' in initializer list [-Wc++11-narrowing]
        float const tmod{fmod(elapsedTime, (float) (((m_loopTime / 202.5f)) / (m_loopTime / 200.f)))};
                         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/wouter/Personal/OpenNFS/src/Camera/HermiteCamera.cpp:13:26: note: insert an explicit cast to silence this issue
        float const tmod{fmod(elapsedTime, (float) (((m_loopTime / 202.5f)) / (m_loopTime / 200.f)))};
                         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                         static_cast<float>(                                                        )
1 error generated.
```

+++ src/Camera/HermiteCamera.cpp
@@ -10,7 +10,7 @@ namespace OpenNFS {
 
     void HermiteCamera::UseSpline(const float elapsedTime) {
         // Ensure we're never sampling the hermite curve outside of points arr size.
-        float const tmod{fmod(elapsedTime, (m_loopTime / 202.5f)) / (m_loopTime / 200.f)};
+        float const tmod{fmodf(elapsedTime, (m_loopTime / 202.5f)) / (m_loopTime / 200.f)};
         position = m_trackCameraRail.GetPointAt(tmod);
 
         // Look towards the position that is a few ms away
